djangoProject/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.3.0
sqlparse==0.4.4
typing-extensions==4.10.0
